Minor configuration fixes

+++ C Programming/Lab 1/config.sh
@@ -1,8 +1,12 @@
-rm -f old_numbers.txt || true
-touch old_numbers.txt
-nano old_numbers.txt
-./a.out
-nano old_numbers.txt
-nano new_numbers.txt
-rm -f new_numbers.txt || true
-rm -f old_numbers.txt || true
+if [ -e "InOut.h" ] && [ -e "maths.h" ] && [ -e "main.c" ]; then
+  gcc -Wall -Werror -std=c99 -pedantic main.c
+  rm -f old_numbers.txt || true
+  touch old_numbers.txt
+  gedit old_numbers.txt
+  ./a.out
+  gedit old_numbers.txt new_numbers.txt
+  rm -f new_numbers.txt || true
+  rm -f old_numbers.txt || true
+else
+  echo 'Some libraries are absent.'
+fi
